+++ CONTRIBUTING.md
@@ -1,0 +1,1 @@
+#Explains how to contribute to the project.
