+++ examples/labels_photos_June08.csv
