--- retrodetect/beetrack.model
@@ -1,105 +1,0 @@
-svm_type c_svc
-kernel_type rbf
-gamma 3.0000000000000001e-05
-nr_class 2
-total_sv 96
-rho -4.3727007484687421
-label 1 0
-nr_sv 47 49
-SV
-300 2:4 3:0.074375004 5:1 6:1 
-189.59450397299773 2:5 3:0.21312501 4:1 5:1 6:-4 
-300 1:1 2:2 3:0.11125 6:4 
-300 1:1 2:3 3:0.0175 6:4 
-300 1:1 2:3 3:0.1425 4:1 6:4 
-300 1:1 2:3 3:0.27250001 5:1 6:7 
-300 1:1 2:4 3:0.260625 5:1 6:11 
-300 1:1 2:4 3:0.391875 4:1 6:9 
-300 1:1 2:4 3:0.49562499 4:1 5:1 6:6 
-300 1:1 2:4 3:0.51187497 4:1 5:1 6:6 
-278.92414443946291 1:1 2:4 3:1.06875 4:1 5:1 6:11 
-300 1:1 2:6 3:0.23625 4:1 5:1 6:11 
-300 1:1 2:7 3:0.174375 4:1 5:1 6:15 
-300 1:1 2:7 3:0.32374999 4:1 5:1 6:17 
-300 1:1 2:7 3:0.82499999 4:1 5:2 6:22 
-211.50163505002826 1:1 2:7 3:1.266875 4:1 5:2 6:23 
-300 1:1 2:12 3:0.63312501 4:3 5:2 6:30 
-300 1:1 2:15 3:0.42750001 4:1 5:5 6:10 
-300 1:2 2:3 3:0.39375001 4:1 5:1 6:9 
-300 1:2 2:3 3:0.44874999 4:1 5:1 6:10 
-300 1:2 2:5 3:0.36375001 4:2 5:1 6:13 
-300 1:2 2:5 3:0.37187499 4:1 5:1 6:13 
-300 1:2 2:5 3:0.40875 4:1 5:1 6:15 
-300 1:2 2:5 3:0.416875 4:1 5:1 6:10 
-300 1:2 2:5 3:0.48124999 4:1 5:1 6:12 
-300 1:2 2:5 3:0.57125002 4:2 5:1 6:4 
-300 1:2 2:5 3:0.98124999 4:2 5:1 6:16 
-300 1:2 2:6 3:0.345 4:1 5:2 6:22 
-300 1:2 2:7 3:0.51999998 4:3 5:1 6:11 
-300 1:2 2:9 3:0.565 4:2 5:2 6:23 
-300 1:2 2:9 3:1.74875 4:2 5:2 6:19 
-300 1:3 2:5 3:1.0443749 4:1 5:3 6:17 
-300 1:3 2:6 3:0.14125 4:1 5:1 6:13 
-300 1:3 2:6 3:0.27125001 4:1 5:1 6:10 
-300 1:3 2:6 3:0.41499999 4:1 5:1 6:9 
-300 1:3 2:6 3:0.52999997 4:1 5:2 6:19 
-300 1:3 2:7 3:0.174375 4:1 5:1 6:15 
-300 1:3 2:7 3:1.89625 4:2 5:3 6:20 
-300 1:3 2:8 3:1.1675 4:2 5:2 6:23 
-300 1:3 2:16 3:0.80687499 4:2 5:2 6:35 
-300 1:3 2:19 3:0.68000001 4:2 5:1 6:37 
-300 1:4 2:5 3:0.43125001 4:3 5:3 6:7 
-168.50734505110745 1:4 2:5 3:0.55000001 4:1 5:4 6:19 
-300 1:5 2:9 3:0.52875 4:3 5:2 6:17 
-300 1:7 2:11 3:0.50749999 4:4 5:3 6:23 
-300 1:19 2:22 3:0.61624998 4:1 5:2 6:48 
-280.53697524523653 1:66 2:70 3:0.91000003 4:5 5:13 6:187 
--300 1:1 2:3 3:0.17812499 4:1 5:1 6:5 
--300 1:1 2:3 3:0.33875 4:1 6:7 
--300 1:1 2:3 3:0.47 4:1 5:1 6:8 
--300 1:1 2:3 3:0.926875 4:1 5:1 6:8 
--300 1:1 2:3 3:0.95249999 4:1 5:1 6:9 
--300 1:1 2:4 3:0.48500001 4:1 5:1 6:7 
--300 1:1 2:4 3:0.700625 4:1 5:1 6:11 
--300 1:1 2:4 3:0.91250002 4:1 5:2 6:9 
--300 1:1 2:4 3:0.995 4:1 5:1 6:13 
--300 1:1 2:8 3:0.301875 4:1 5:1 6:15 
--300 1:1 2:10 3:0.51625001 4:1 5:1 6:17 
--300 1:1 2:17 3:0.59562498 4:2 5:1 6:34 
--300 1:1 2:17 3:0.59937501 4:1 5:2 6:34 
--300 1:2 2:3 3:0.33875 4:1 6:7 
--300 1:2 2:5 3:0.19750001 4:1 5:1 6:11 
--300 1:2 2:5 3:0.36625001 4:1 5:1 6:10 
--300 1:2 2:5 3:0.703125 4:1 5:1 6:12 
--300 1:2 2:5 3:0.78187501 4:1 5:1 6:10 
--300 1:2 2:8 3:0.28874999 5:1 6:11 
--300 1:3 2:4 3:0.030625001 5:1 6:11 
--300 1:3 2:4 3:0.41312501 4:1 5:1 6:10 
--300 1:3 2:5 3:0.0093750004 5:1 6:7 
--300 1:3 2:5 3:0.609375 4:1 5:1 6:12 
--300 1:3 2:5 3:1.04625 4:1 5:2 6:13 
--300 1:3 2:6 3:0.293125 4:1 5:1 6:12 
--300 1:3 2:6 3:0.70875001 4:1 5:1 6:16 
--300 1:3 2:6 3:0.95375001 4:1 5:1 6:12 
--300 1:3 2:6 3:0.95687503 4:1 5:1 6:12 
--300 1:3 2:6 3:1.0325 4:2 5:2 6:13 
--300 1:3 2:10 3:0.215 5:1 6:19 
--300 1:3 2:15 3:0.52999997 4:2 5:1 6:32 
--300 1:3 2:17 3:0.61812502 4:1 5:2 6:37 
--300 1:4 2:6 3:0.101875 6:9 
--300 1:4 2:7 3:0.51499999 4:1 5:1 6:13 
--300 1:4 2:8 3:0.53312498 4:1 5:1 6:15 
--300 1:4 2:8 3:0.64187503 4:1 5:1 6:15 
--300 1:4 2:8 3:0.99437499 4:1 5:1 6:19 
--300 1:4 2:9 3:1.00625 4:1 5:1 6:16 
--300 1:4 2:10 3:0.48374999 4:1 5:1 6:22 
--300 1:5 2:7 3:0.451875 4:1 5:1 6:13 
--300 1:5 2:11 3:0.59937501 4:1 5:2 6:23 
--223.58634725222572 1:6 2:7 3:0.37562501 4:1 5:1 6:16 
--300 1:6 2:8 3:0.96812499 4:1 5:2 6:20 
--300 1:6 2:8 3:0.97437501 4:1 5:2 6:15 
--300 1:6 2:9 3:0.98562503 4:2 5:2 6:17 
--0.13701620126587777 1:44 2:136 3:1.14625 4:2 5:10 6:271 
--292.65270339446238 1:61 2:68 3:1.71 4:3 5:5 6:184 
--8.9475187687180107 1:124 2:130 3:1.04375 4:2 5:6 6:273 
--3.741018142161197 1:249 2:255 3:0.78187501 4:1 5:5 6:373
