+++ src/bee-retrodetect/normxcorr2.py
